Refactor RPC API: Modularize file and staking operations
- Moved file upload and retrieval functions to a new `metadata` module for better organization.
- Created a new `stake` module to encapsulate staking-related functions, including staking, unstaking, and retrieving staking information.
- Removed redundant code from the main RPC API file, improving readability and maintainability.
- Updated the RPC server to utilize the new modular structure for file and staking operations.
