+++ Cargo.toml
@@ -105,7 +105,7 @@ jsonrpc-http-server = "18.0"
 tokio = { version = "1", features = ["full"] }
 url = "2.5.4"
 reqwest = { version = "0.12.15"}
-
+base64 = "0.22.1"
 # Error Handling & Diagnostics
 anyhow = "1.0.95"
 codespan-reporting = "0.11.0"
