+++ crates/rpc/rpc-api/Cargo.toml
@@ -12,22 +12,26 @@ repository.workspace = true
 description.workspace = true
 
 [dependencies]
-# Networking
+# Existing dependencies
 jsonrpc-core.workspace = true
 jsonrpc-http-server.workspace = true
-futures.workspace = true
-tokio.workspace = true
-
-
-# Serialization
-serde_json.workspace = true
 serde.workspace = true
-
-network.workspace = true
-consensus-pos.workspace = true
-panorama.workspace = true
-key.workspace = true
-mona-storage = { path = "../../../monaos/mona-storage" }
-mona-types = { path = "../../../monaos/mona-types" }
-base64 = "0.22.1"
+serde_json.workspace = true
+futures.workspace = true
+log.workspace = true
+base64.workspace = true
 chrono.workspace = true
+hex.workspace = true
+common.workspace = true
+
+# Dependencies for HTTPS support
+rustls = "0.21.5"
+rustls-pemfile = "1.0.2"
+
+# Project dependencies
+network = { workspace = true }
+panorama = { workspace = true }
+mona-types = { workspace = true }
+mona-storage = { workspace = true }
+key = { workspace = true }
+tokio = { workspace = true, features = ["full"] }
